pyproject.toml
[tool.poetry]
name = "database_tests_with_sqlalchemy"
version = "0.1.0"
description = "Database Tests with SQLAlchemy"
authors = ["Serhii Smetanskyi"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.2"
pydantic-settings = "^2.0.1"
alembic = "^1.7.4"
sqlalchemy = "^2.0.19"
psycopg2-binary = ">=2.9.7"
python-dotenv = ">=1.0.0"
passlib = "^1.7.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
setuptools = "^68.2.2"
black = "^23.11.0"
pylint = "^2.17.4"
mypy = "^1.4.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.3.2"
factory-boy = "^3.2.1"
faker = "^19.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

